requirements.txt
flask==3.0.0
requests==2.31.0
google-generativeai==0.3.1
python-dotenv==1.0.0
PyGithub==2.1.1
cryptography==41.0.7

